Move SymbolReaderHttpHandler and associated interfaces to new project.
