+++ src/PerfView/SymbolReaderHttpHandler.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation.  All rights reserved
 using Azure.Core;
 using Azure.Identity;
-using Microsoft.IdentityModel.JsonWebTokens;
+using Microsoft.Diagnostics.Symbols.Authentication;
 using PerfView.Dialogs;
 using System;
 using System.Collections.Concurrent;
@@ -21,733 +21,6 @@ using System.Windows.Interop;
 
 namespace PerfView
 {
-    /// <summary>
-    /// A handler for <see cref="HttpClient"/> that allows interception and
-    /// modification of requests by a chain of <see cref="IAsyncSymbolReaderHandler"/>
-    /// instances.
-    /// </summary>
-    /// <remarks>
-    /// In PerfView, this is usually passed to the constructor of
-    /// <see cref="Microsoft.Diagnostics.Symbols.SymbolReader"/>.
-    /// </remarks>
-    internal class SymbolReaderHttpHandler : DelegatingHandler
-    {
-        /// <summary>
-        /// A list of handlers.
-        /// </summary>
-        private readonly List<IAsyncSymbolReaderHandler> _handlers = new List<IAsyncSymbolReaderHandler>();
-
-        /// <summary>
-        /// Construct a new <see cref="SymbolReaderHttpHandler"/> instance.
-        /// </summary>
-        public SymbolReaderHttpHandler() : base(new HttpClientHandler() { CheckCertificateRevocationList = true })
-        {
-        }
-
-        /// <summary>
-        /// Gets or sets whether this instance has been disposed.
-        /// </summary>
-        public bool IsDisposed { get; private set; }
-
-        /// <inheritdoc/>
-        protected override void Dispose(bool disposing)
-        {
-            IsDisposed = true;
-            base.Dispose(disposing);
-        }
-
-        /// <summary>
-        /// Process the request by forwarding it to all the handlers.
-        /// </summary>
-        /// <param name="request">The request.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>The response.</returns>
-        protected async override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-        {
-            var context = new RequestContext(request);
-
-            // Build a delegate chain, starting with the root handler.
-            SymbolReaderHandlerDelegate chain = () => RootHandlerAsync(context, cancellationToken);
-
-            // Prevent modification while we build the chain.
-            lock (_handlers)
-            {
-                // Loop through the handlers in reverse so that the resulting
-                // chain is in the order the handlers were added.
-                for (int i = _handlers.Count - 1; i >= 0; i--)
-                {
-                    IAsyncSymbolReaderHandler handler = _handlers[i];
-                    var next = new SymbolReaderHandlerDelegate(chain);
-                    chain = () => handler.ProcessRequestAsync(context, next, cancellationToken);
-                }
-            }
-
-            await chain().ConfigureAwait(false);
-            return context.Response;
-        }
-
-        /// <summary>
-        /// The root of the delegate chain.
-        /// Calls the inner handler and sets the response
-        /// on the <see cref="RequestContext"/> object.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>A task that completes when the response has been set.</returns>
-        private async Task RootHandlerAsync(RequestContext context, CancellationToken cancellationToken)
-        {
-            context.Response = await base.SendAsync(context.Request, cancellationToken).ConfigureAwait(false);
-        }
-
-        /// <summary>
-        /// Remove all the handlers.
-        /// </summary>
-        public void ClearHandlers()
-        {
-            lock (_handlers)
-            {
-                foreach (var handler in _handlers)
-                {
-                    if (handler is IDisposable disposable)
-                    {
-                        disposable.Dispose();
-                    }
-                }
-
-                _handlers.Clear();
-            }
-        }
-
-        /// <summary>
-        /// Add a handler. Handlers will be called in the order they are added.
-        /// </summary>
-        /// <param name="handler">The handler to add.</param>
-        /// <returns>This instance for fluent chaining.</returns>
-        public SymbolReaderHttpHandler AddHandler(IAsyncSymbolReaderHandler handler)
-        {
-            lock (_handlers)
-            {
-                _handlers.Add(handler);
-            }
-
-            return this;
-        }
-
-        /// <summary>
-        /// Add a handler that uses Git Credential Manager for authentication.
-        /// </summary>
-        /// <param name="log">A logger.</param>
-        /// <param name="mainWindow">The main window to use when parenting modal dialogs.</param>
-        /// <returns>This instance for fluent chaining.</returns>
-        public SymbolReaderHttpHandler AddGitCredentialManagerAuthentication(TextWriter log, Window mainWindow)
-            => AddHandler(new GitCredentialManagerHandler(log, GetWindowHandle(mainWindow)));
-
-        /// <summary>
-        /// Add a handler for Azure DevOps authentication using local credentials.
-        /// It will try to use cached credentials from Visual Studio, VS Code, 
-        /// Azure Powershell and Azure CLI.
-        /// </summary>
-        /// <param name="log">A logger.</param>
-        /// <param name="silent">If no local credentials can be found, then a browser window will
-        /// be opened to prompt the user. Set this to true to if you don't want that.</param>
-        /// <returns>This instance for fluent chaining.</returns>
-        public SymbolReaderHttpHandler AddAzureDevOpsAuthentication(TextWriter log, bool silent = false)
-        {
-            DefaultAzureCredentialOptions options = new DefaultAzureCredentialOptions
-            {
-                ExcludeInteractiveBrowserCredential = silent,
-                ExcludeManagedIdentityCredential = true // This is not designed to be used in a service.
-            };
-
-            return AddHandler(new AzureDevOpsHandler(log, new DefaultAzureCredential(options)));
-        }
-
-        /// <summary>
-        /// Add a handler for Symweb authentication using local credentials.
-        /// It will try to use cached credentials from Visual Studio, VS Code,
-        /// Azure Powershell and Azure CLI.
-        /// </summary>
-        /// <param name="log">A logger.</param>
-        /// <param name="silent">If no local credentials can be found, then a browser window will
-        /// be opened to prompt the user. Set this to true to if you don't want that.</param>
-        /// <returns>This instance for fluent chaining.</returns>
-        public SymbolReaderHttpHandler AddSymwebAuthentication(TextWriter log, bool silent = false)
-        {
-            DefaultAzureCredentialOptions options = new DefaultAzureCredentialOptions
-            {
-                ExcludeInteractiveBrowserCredential = silent,
-                ExcludeManagedIdentityCredential = true // This is not designed to be used in a service.
-            };
-
-            return AddHandler(new SymwebHandler(log, new DefaultAzureCredential(options)));
-        }
-
-        /// <summary>
-        /// Add a handler for GitHub device flow authentication.
-        /// </summary>
-        /// <param name="log">A logger.</param>
-        /// <param name="mainWindow">The Window to use for parenting any modal
-        /// dialogs needed for authentication.</param>
-        /// <returns>This instance for fluent chaining.</returns>
-        public SymbolReaderHttpHandler AddGitHubDeviceCodeAuthentication(TextWriter log, Window mainWindow)
-            => AddHandler(new GitHubDeviceFlowHandler(log, mainWindow));
-
-        public SymbolReaderHttpHandler AddBasicHttpAuthentication(TextWriter log, Window mainWindow)
-            => AddHandler(new BasicHttpAuthHandler(log));
-
-        /// <summary>
-        /// Get the HWND of the given WPF window in a way that honors WPF
-        /// threading rules.
-        /// </summary>
-        /// <param name="window">The WPF window.</param>
-        /// <returns>The handle (HWND) of the given window.</returns>
-        private static IntPtr GetWindowHandle(Window window)
-            => window.Dispatcher.Invoke(() => new WindowInteropHelper(window).Handle);
-    }
-
-    /// <summary>
-    /// The delegate type that handlers invoke to chain to the next handler.
-    /// </summary>
-    /// <returns>A task.</returns>
-    internal delegate Task SymbolReaderHandlerDelegate();
-
-    /// <summary>
-    /// Context object passed to <see cref="IAsyncSymbolReaderHandler.ProcessRequestAsync(RequestContext, SymbolReaderHandlerDelegate, CancellationToken)"/>
-    /// </summary>
-    internal sealed class RequestContext
-    {
-        /// <summary>
-        /// The response.
-        /// </summary>
-        private HttpResponseMessage _response;
-
-        public RequestContext(HttpRequestMessage request)
-        {
-            Request = request;
-        }
-
-        /// <summary>
-        /// The request. Handlers may modify this as necessary
-        /// before or after calling the next handler in the chain.
-        /// </summary>
-        public HttpRequestMessage Request { get; }
-
-        /// <summary>
-        /// Gets or sets the response.It is initially null, and
-        /// it will be set by the innermost handler. Handlers
-        /// may examine it and modify it, if necessary, after
-        /// calling the next handler in the chain.
-        /// </summary>
-        public HttpResponseMessage Response
-        {
-            get => _response;
-            set
-            {
-                if (value != _response)
-                {
-                    if (_response != null)
-                    {
-                        _response.Dispose();
-                    }
-
-                    _response = value;
-                }
-            }
-        }
-    }
-
-    /// <summary>
-    /// Enumeration of supported authentication schemes.
-    /// </summary>
-    internal enum AuthScheme
-    {
-        /// <summary>
-        /// Basic authentication.
-        /// </summary>
-        Basic,
-
-        /// <summary>
-        /// Bearer token.
-        /// </summary>
-        Bearer
-    }
-
-    /// <summary>
-    /// This is similar to <see cref="AccessToken"/>, but it includes
-    /// the authentication scheme.
-    /// </summary>
-    internal readonly struct AuthToken : IEquatable<AuthToken>
-    {
-        /// <summary>
-        /// Creates a new instance of <see cref="AuthToken"/> using the supplied
-        /// parameters.
-        /// </summary>
-        /// <param name="scheme">The scheme.</param>
-        /// <param name="token">The token value.</param>
-        /// <param name="expiresOn">The token's expiry time.</param>
-        /// <param name="userName">Optional user name.</param>
-        public AuthToken(AuthScheme scheme, string token, DateTime expiresOn, string userName = null)
-        {
-            Scheme = scheme;
-            Token = token;
-            ExpiresOn = expiresOn;
-            UserName = userName;
-            Password = null;
-        }
-
-        /// <summary>
-        /// Create a <see cref="AuthScheme.Basic"/> using the given username
-        /// and password.
-        /// </summary>
-        /// <param name="username">The username.</param>
-        /// <param name="password">The password.</param>
-        private AuthToken(string username, string password)
-        {
-            Scheme = AuthScheme.Basic;
-            UserName = username;
-            Password = password;
-            ExpiresOn = DateTime.UtcNow + TimeSpan.FromDays(60);
-
-            // The Token is the Base-64-encoded value representing the
-            // username and password combined.
-            // The lifetime is set to 60 days.
-            string secret = username + ':' + password;
-            byte[] utf8Secret = Encoding.UTF8.GetBytes(secret);
-            Token = Convert.ToBase64String(utf8Secret);
-        }
-
-        /// <summary>
-        /// Creates a new <see cref="AuthToken"/> with the <see cref="AuthScheme.Basic"/> 
-        /// scheme from a user name and password. A single token value is computed as a
-        /// Base-64 encoded value representing the username and password combined.
-        /// If no lifetime is specified, the token will be valid for 60 days.
-        /// </summary>
-        /// <param name="username">The user name.</param>
-        /// <param name="password">The password.</param>
-        /// <returns>A constructed <see cref="AuthToken"/>.</returns>
-        public static AuthToken CreateBasicFromUsernameAndPassword(string username, string password)
-        {
-            if (string.IsNullOrEmpty(username))
-            {
-                throw new ArgumentException($"'{nameof(username)}' cannot be null or empty.", nameof(username));
-            }
-
-            if (string.IsNullOrEmpty(password))
-            {
-                throw new ArgumentException($"'{nameof(password)}' cannot be null or empty.", nameof(password));
-            }
-
-            return new AuthToken(username, password);
-        }
-
-        /// <summary>
-        /// Try to parse an <see cref="AuthToken"/> from an OAuth bearer token
-        /// as a Json Web Token.
-        /// </summary>
-        /// <param name="userName">The user name, if any.</param>
-        /// <param name="token">The Json Web Token (JWT) value.</param>
-        /// <param name="authToken">The auth token.</param>
-        /// <returns>True if the token could be parsed.</returns>
-        public static bool TryParseFromOAuthToken(string userName, string token, out AuthToken authToken)
-        {
-            try
-            {
-                var jwt = new JsonWebToken(token);
-                authToken= new AuthToken(AuthScheme.Bearer, token, jwt.ValidTo, userName);
-                return true;
-            }
-            catch (Exception)
-            {
-                authToken = default;
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Creates a new instance of <see cref="AuthToken"/> from an <see cref="AccessToken"/>
-        /// using the <see cref="AuthScheme.Bearer"/> scheme.
-        /// </summary>
-        /// <param name="token">The access token from Azure.Core</param>
-        public static AuthToken FromAzureCoreAccessToken(AccessToken token)
-            => new AuthToken(AuthScheme.Bearer, token.Token, token.ExpiresOn.UtcDateTime);
-
-        /// <summary>
-        /// Gets the authentication scheme to use with this token.
-        /// </summary>
-        public AuthScheme Scheme { get; }
-
-        /// <summary>
-        /// Gets the value of the token.
-        /// </summary>
-        public string Token { get; }
-
-        /// <summary>
-        /// Gets the time when the token expires.
-        /// </summary>
-        public DateTime ExpiresOn { get; }
-
-        /// <summary>
-        /// The username. May be null.
-        /// </summary>
-        public string UserName { get; }
-
-        /// <summary>
-        /// The password part of a basic authentication token.
-        /// </summary>
-        public string Password { get; }
-
-        public override bool Equals(object obj) => obj is AuthToken token && Equals(token);
-
-        public bool Equals(AuthToken other)
-        {
-            // Only the Scheme and Token take part
-            // in equality.
-            return Scheme == other.Scheme &&
-                   Token == other.Token;
-        }
-
-        public override int GetHashCode()
-        {
-            // Only the Scheme and Token take part
-            // in equality.
-            int hashCode = -41169086;
-            hashCode = hashCode * -1521134295 + Scheme.GetHashCode();
-            hashCode = hashCode * -1521134295 + StringComparer.Ordinal.GetHashCode(Token);
-            return hashCode;
-        }
-
-        public static bool operator ==(AuthToken left, AuthToken right) => left.Equals(right);
-
-        public static bool operator !=(AuthToken left, AuthToken right) => !(left == right);
-    }
-
-    /// <summary>
-    /// Interface for asynchronous symbol reader handlers.
-    /// </summary>
-    internal interface IAsyncSymbolReaderHandler
-    {
-        /// <summary>
-        /// Process the given request, modifying it if necessary and passing it on to the next handler.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="next">The next handler in the chain.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>A task that completes when desired modifications have been made.</returns>
-        Task ProcessRequestAsync(RequestContext context, SymbolReaderHandlerDelegate next, CancellationToken cancellationToken);
-    }
-
-    /// <summary>
-    /// Common base class for all <see cref="IAsyncSymbolReaderHandler"/> implementations.
-    /// Adds a logger with a handler-specific prefix.
-    /// </summary>
-    internal abstract class SymbolReaderHandlerBase : IAsyncSymbolReaderHandler
-    {
-        /// <summary>
-        /// The logger.
-        /// </summary>
-        private readonly TextWriter _log;
-
-        /// <summary>
-        /// A prefix for all messages written by <see cref="WriteLog(string, object[])"/>.
-        /// </summary>
-        private readonly string _logPrefix;
-
-        /// <summary>
-        /// Creates a new <see cref="SymbolReaderHandlerBase"/> with the given parameters.
-        /// </summary>
-        /// <param name="log">A logger. May be null.</param>
-        /// <param name="logPrefix">A prefix for log messages. May be null.</param>
-        protected SymbolReaderHandlerBase(TextWriter log, string logPrefix)
-        {
-            _log = log ?? TextWriter.Null;
-            _logPrefix = logPrefix;
-        }
-
-        /// <summary>
-        /// Implemented in derived classes to handle the requests.
-        /// Implementations should process the given request, modifying it if necessary
-        /// before calling the next handler.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="next">The next handler in the chain.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>A task that completes when desired modifications have been made.</returns>
-        public abstract Task ProcessRequestAsync(RequestContext context, SymbolReaderHandlerDelegate next, CancellationToken cancellationToken);
-
-        /// <summary>
-        /// Write a line to the log, prefixed with <see cref="_logPrefix"/>.
-        /// </summary>
-        /// <param name="format">A composite format string with optional placeholders.</param>
-        /// <param name="args">An object array that contains zero or more objects to format and write.</param>
-        protected void WriteLog(string format, params object[] args)
-        {
-            if (!string.IsNullOrEmpty(_logPrefix))
-            {
-                _log.Write(_logPrefix);
-            }
-
-            _log.WriteLine(format, args);
-        }
-
-        /// <summary>
-        /// Write a line to the log file and enclose it in square brackets
-        /// so that the status bar logger will pick it up.
-        /// </summary>
-        /// <param name="format">A composite format string with optional placeholders.</param>
-        /// <param name="args">An object array that contains zero or more objects to format and write.</param>
-        protected void WriteStatusLog(string format, params object[] args)
-        {
-            if (!string.IsNullOrEmpty(_logPrefix))
-            {
-                _log.Write(_logPrefix);
-            }
-
-            _log.WriteLine('[' + string.Format(format, args) + ']');
-        }
-    }
-
-    /// <summary>
-    /// Base class for request handlers that add authorization headers to
-    /// requests.
-    /// </summary>
-    internal abstract class SymbolReaderAuthHandlerBase : SymbolReaderHandlerBase
-    {
-        /// <summary>
-        /// The scheme name for Bearer tokens.
-        /// </summary>
-        protected const string BearerScheme = "Bearer";
-
-        /// <summary>
-        /// The scheme for Basic tokens.
-        /// </summary>
-        private const string BasicScheme = "Basic";
-
-        /// <summary>
-        /// A cache of access tokens per authority. It is shared by all
-        /// <see cref="SymbolReaderAuthHandlerBase"/> implementations.
-        /// </summary>
-        private static readonly ConcurrentDictionary<Uri, AuthToken> s_tokenCache = new ConcurrentDictionary<Uri, AuthToken>();
-
-        /// <summary>
-        /// Constructor to be called from derived classes.
-        /// </summary>
-        /// <param name="log">A logger. May be null.</param>
-        /// <param name="logPrefix">A prefix for log messages. May be null.</param>
-        protected SymbolReaderAuthHandlerBase(TextWriter log, string logPrefix) : base(log, logPrefix)
-        {
-        }
-
-        /// <summary>
-        /// Handles the request and calls the next handler in the chain, if necessary.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="next">The next handler.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>A task that completes when the request has been handled.</returns>
-        public override Task ProcessRequestAsync(RequestContext context, SymbolReaderHandlerDelegate next, CancellationToken cancellationToken)
-        {
-            HttpRequestMessage request = context.Request;
-            if (request.RequestUri.Scheme == Uri.UriSchemeHttps &&  // Require a secure connection
-                request.Headers.Authorization is null &&            // No existing auth header
-                TryGetAuthority(request.RequestUri, out Uri authority) // Is a recognized authority
-                )
-            {
-                if (!TryGetCachedToken(authority, logExpiration: true, out AuthToken cachedToken))
-                {
-                    return GetAuthTokenCoreAsync(context, next, authority, cancellationToken);
-                }
-
-                AddAuthorizationHeader(request, cachedToken);
-            }
-
-            return next();
-        }
-
-        /// <summary>
-        /// Call the derived class to get an access token for the given authority.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="next">The next handler.</param>
-        /// <param name="authority">The authority to use to acquire a new token.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns></returns>
-        private async Task GetAuthTokenCoreAsync(RequestContext context, SymbolReaderHandlerDelegate next, Uri authority, CancellationToken cancellationToken)
-        {
-            AuthToken? token = await GetAuthTokenAsync(context, next, authority, cancellationToken).ConfigureAwait(false);
-            if (!token.HasValue)
-            {
-                // Call the next handler only if the derived class didn't.
-                if (context.Response is null)
-                {
-                    await next().ConfigureAwait(false);
-                }
-
-                return;
-            }
-
-            TryAddCachedToken(authority, token.Value);
-            AddAuthorizationHeader(context.Request, token.Value);
-
-            await next().ConfigureAwait(false);
-
-            if (context.Response?.StatusCode == HttpStatusCode.Unauthorized)
-            {
-                s_tokenCache.TryRemove(authority, out _);
-                await RejectTokenAsync(context, authority, token.Value, cancellationToken).ConfigureAwait(false);
-            }
-            else
-            {
-                await AcceptTokenAsync(context, authority, token.Value, cancellationToken).ConfigureAwait(false);
-            }
-        }
-
-        /// <summary>
-        /// May be overridden in derived classes to handle the case when 
-        /// authorization with a token succeeded.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="authority">The authority.</param>
-        /// <param name="token">The bad token.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>A task that completes when token acceptance has finished.</returns>
-        protected virtual Task AcceptTokenAsync(RequestContext context, Uri authority, AuthToken token, CancellationToken cancellationToken) => Task.CompletedTask;
-
-        /// <summary>
-        /// May be overridden in derived classes to handle the case when 
-        /// authorization failed.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="authority">The authority.</param>
-        /// <param name="token">The bad token.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>A task that completes when token rejection has finished.</returns>
-        protected virtual Task RejectTokenAsync(RequestContext context, Uri authority, AuthToken token, CancellationToken cancellationToken) => Task.CompletedTask;
-
-        /// <summary>
-        /// Implemented in derived classes to try and determine the authority from the request URI.
-        /// </summary>
-        /// <param name="requestUri">The request URI.</param>
-        /// <param name="authority">The authority, if known.</param>
-        /// <returns>True if <paramref name="requestUri"/> is one you recognize and know the authority.</returns>
-        protected abstract bool TryGetAuthority(Uri requestUri, out Uri authority);
-
-        /// <summary>
-        /// Overridden in derived classes to get an access token for the given authority.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="next">The next handler. Derived implementations do not need to
-        /// call this unless they need to handle the unauthenticated response. The base
-        /// class will check if <see cref="RequestContext.Response"/> is already set and
-        /// stop further processing.</param>
-        /// <param name="authority">The authority.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>An access token, if available, or null otherwise.</returns>
-        protected virtual Task<AuthToken?> GetAuthTokenAsync(RequestContext context, SymbolReaderHandlerDelegate next, Uri authority, CancellationToken cancellationToken) => Task.FromResult<AuthToken?>(null);
-
-        /// <summary>
-        /// Add an authorization header to the request using the Bearer scheme.
-        /// </summary>
-        /// <param name="request">The request to modify.</param>
-        /// <param name="authToken">The authorization token to add.</param>
-        protected static void AddAuthorizationHeader(HttpRequestMessage request, AuthToken authToken)
-        {
-            string scheme = GetAuthenticationHeaderScheme(authToken);
-            request.Headers.Authorization = new AuthenticationHeaderValue(scheme, authToken.Token);
-        }
-
-        /// <summary>
-        /// Checks if the access token is still valid vis-à-vis its lifetime.
-        /// </summary>
-        /// <param name="token">The access token.</param>
-        /// <returns><c>true</c> if the access token has not expired and is not
-        /// going to expire soon.</returns>
-        protected static bool IsAccessTokenStillGood(in AuthToken token)
-        {
-            if (token.ExpiresOn == default || token.Token == null)
-            {
-                return false;
-            }
-
-            TimeSpan timeLeft = token.ExpiresOn - DateTime.UtcNow;
-
-            // Try to refresh at least 5 minutes before expiration.
-            return timeLeft > TimeSpan.FromMinutes(5);
-        }
-
-        /// <summary>
-        /// Derived classes may call this if they need to cache additional access tokens
-        /// for authorities other than the one returned by <see cref="TryGetAuthority(Uri, out Uri)"/> 
-        /// (which are already cached by <see cref="SymbolReaderAuthHandlerBase"/>).
-        /// </summary>
-        /// <param name="authority">The authority.</param>
-        /// <param name="token">The token.</param>
-        /// <returns>True if the token was successfully cached.</returns>
-        protected bool TryAddCachedToken(Uri authority, in AuthToken token) => s_tokenCache.TryAdd(authority, token);
-
-        /// <summary>
-        /// Retrieve a token from the cache. If a cached token has expired, or is close to
-        /// the end of its lifetime, then it will be removed from the cache and this
-        /// will return false.
-        /// </summary>
-        /// <param name="authority">The authority.</param>
-        /// <param name="token">The token.</param>
-        /// <returns>True if a good access token was found in the cache.</returns>
-        protected bool TryGetCachedToken(Uri authority, out AuthToken token) => TryGetCachedToken(authority, logExpiration: false, out token);
-
-        /// <summary>
-        /// Retrieve a token from the cache. If a cached token has expired, or is close to
-        /// the end of its lifetime, then it will be removed from the cache and this
-        /// will return false.
-        /// </summary>
-        /// <param name="authority">The authority.</param>
-        /// <param name="logExpiration">Whether to write a line to the log when an expired token is removed from the cache.</param>
-        /// <param name="token">The token.</param>
-        /// <returns>True if a good access token was found in the cache.</returns>
-        private bool TryGetCachedToken(Uri authority, bool logExpiration, out AuthToken token)
-        {
-            if (s_tokenCache.TryGetValue(authority, out token))
-            {
-                if (IsAccessTokenStillGood(token))
-                {
-                    return true;
-                }
-
-                if (logExpiration)
-                {
-                    WriteLog("The authorization token for {0} has expired (or is close to expiration).");
-                }
-
-                s_tokenCache.TryRemove(authority, out _);
-            }
-
-            token = default;
-            return false;
-        }
-
-        /// <summary>
-        /// Examines the <see cref="AuthToken.Scheme"/> value of an
-        /// <see cref="AuthToken"/> and converts it to a string suitable for
-        /// an HTTP Authorization header.
-        /// </summary>
-        /// <param name="authToken">The authorization token.</param>
-        /// <returns>A string version of the scheme.</returns>
-        /// <exception cref="ArgumentException">The scheme is unrecognized.</exception>
-        private static string GetAuthenticationHeaderScheme(AuthToken authToken)
-        {
-            switch (authToken.Scheme)
-            {
-                case AuthScheme.Bearer:
-                    return BearerScheme;
-
-                case AuthScheme.Basic:
-                    return BasicScheme;
-
-                default:
-                    throw new ArgumentException($"'{nameof(authToken)}' specified an unrecognized {nameof(AuthToken.Scheme)}", nameof(authToken));
-            }
-        }
-    }
-
     /// <summary>
     /// A handler that adds support for basic username:password authentication over HTTP
     /// </summary>
@@ -1593,101 +866,6 @@ namespace PerfView
     }
 
     /// <summary>
-    /// A handler that adds authorization for Symweb.
-    /// </summary>
-    internal sealed class SymwebHandler : SymbolReaderAuthHandlerBase
-    {
-        /// <summary>
-        /// The value of <see cref="Symweb.Scope"/> stored in a single element
-        /// array suitable for passing to
-        /// <see cref="TokenCredential.GetTokenAsync(TokenRequestContext, CancellationToken)"/>.
-        /// </summary>
-        private static readonly string[] s_scopes = new[] { Symweb.Scope };
-
-        /// <summary>
-        /// Prefix to put in front of logging messages.
-        /// </summary>
-        private const string LogPrefix = "SymwebAuth: ";
-
-        /// <summary>
-        /// A provider of access tokens.
-        /// </summary>
-        private readonly TokenCredential _tokenCredential;
-
-        /// <summary>
-        /// Protect <see cref="_tokenCredential"/> against concurrent access.
-        /// </summary>
-        private readonly SemaphoreSlim _tokenCredentialGate = new SemaphoreSlim(initialCount: 1);
-
-        /// <summary>
-        /// An HTTP client used to discover the authority (login endpoint and tenant) for Symweb.
-        /// </summary>
-        private readonly HttpClient _httpClient = new HttpClient(new HttpClientHandler() { CheckCertificateRevocationList = true });
-
-        /// <summary>
-        /// Construct a new <see cref="SymwebHandler"/> instance.
-        /// </summary>
-        /// <param name="tokenCredential">A provider of access tokens.</param>
-        public SymwebHandler(TextWriter log, TokenCredential tokenCredential) : base(log, LogPrefix)
-        {
-            _tokenCredential = tokenCredential ?? throw new ArgumentNullException(nameof(tokenCredential));
-        }
-
-        /// <summary>
-        /// Try to find the authority endpoint for Symweb
-        /// given a full URI.
-        /// </summary>
-        /// <param name="requestUri">The request URI.</param>
-        /// <param name="authority">The authority, if found.</param>
-        /// <returns>True if <paramref name="requestUri"/> represents a path to a
-        /// resource in Symweb.</returns>
-        protected override bool TryGetAuthority(Uri requestUri, out Uri authority) => Symweb.TryGetAuthority(requestUri, out authority);
-
-        /// <summary>
-        /// Get a token to access Symweb.
-        /// </summary>
-        /// <param name="context">The request context.</param>
-        /// <param name="next">The next handler.</param>
-        /// <param name="authority">The Symweb instance.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>An access token, or null if one could not be obtained.</returns>
-        protected override async Task<AuthToken?> GetAuthTokenAsync(RequestContext context, SymbolReaderHandlerDelegate next, Uri authority, CancellationToken cancellationToken)
-        {
-            // Get a new access token from the credential provider.
-            WriteLog("Asking for authorization to access {0}", authority);
-            AuthToken token = await GetTokenAsync(cancellationToken).ConfigureAwait(false);
-
-            return token;
-        }
-
-        /// <summary>
-        /// Get a new access token for Symweb from the <see cref="TokenCredential"/>.
-        /// </summary>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>The access token.</returns>
-        private async Task<AuthToken> GetTokenAsync(CancellationToken cancellationToken)
-        {
-            await _tokenCredentialGate.WaitAsync(cancellationToken).ConfigureAwait(false);
-            try
-            {
-                // Use the token credential provider to acquire a new token.
-                TokenRequestContext requestContext = new TokenRequestContext(s_scopes);
-                AccessToken accessToken = await _tokenCredential.GetTokenAsync(requestContext, cancellationToken).ConfigureAwait(false);
-                return AuthToken.FromAzureCoreAccessToken(accessToken);
-            }
-            catch (Exception ex)
-            {
-                WriteStatusLog("Exception getting token. {0}", ex);
-                throw;
-            }
-            finally
-            {
-                _tokenCredentialGate.Release();
-            }
-        }
-    }
-
-    /// <summary>
     /// A handler that handles GitHub device flow authorization.
     /// </summary>
     internal sealed class GitHubDeviceFlowHandler : SymbolReaderAuthHandlerBase
@@ -1973,64 +1151,6 @@ namespace PerfView
             /// </summary>
             [JsonPropertyName("scope")]
             public string Scope { get; set; }
-        }
-    }
-
-    /// <summary>
-    /// Contains constants, static properties and helper methods pertinent to Symweb.
-    /// </summary>
-    internal static class Symweb
-    {
-        /// <summary>
-        /// The OAuth scope to use when requesting tokens for Symweb.
-        /// </summary>
-        public const string Scope = "af9e1c69-e5e9-4331-8cc5-cdf93d57bafa/.default";
-
-        /// <summary>
-        /// The url host for Symweb.
-        /// </summary>
-        public const string SymwebHost = "symweb.azurefd.net";
-
-        /// <summary>
-        /// Try to find the authority endpoint for Symweb given a full URI.
-        /// </summary>
-        /// <param name="requestUri">The request URI.</param>
-        /// <param name="authority">The authority, if found.</param>
-        /// <returns>True if <paramref name="requestUri"/> represents a path to a
-        /// resource in Symweb.</returns>
-        public static bool TryGetAuthority(Uri requestUri, out Uri authority)
-        {
-            if (!requestUri.IsAbsoluteUri)
-            {
-                authority = null;
-                return false;
-            }
-
-            UriBuilder builder = null;
-            string host = requestUri.DnsSafeHost;
-            if (host.Equals(SymwebHost, StringComparison.OrdinalIgnoreCase))
-            {
-                    builder = new UriBuilder
-                    {
-                        Host = SymwebHost
-                    };
-            }
-
-            if (builder is null)
-            {
-                // Not a Symweb URI.
-                authority = null;
-                return false;
-            }
-
-            builder.Scheme = requestUri.Scheme;
-            if (!requestUri.IsDefaultPort)
-            {
-                builder.Port = requestUri.Port;
-            }
-
-            authority = builder.Uri;
-            return true;
         }
     }
